Moves + renames.
