+++ test/Autofac.Pooling.Tests/DisposableTests.cs
@@ -1,5 +1,5 @@
 ﻿using Autofac.Features.OwnedInstances;
-using Autofac.Pooling.Tests.Types;
+using Autofac.Pooling.Tests.Shared;
 using Xunit;
 
 namespace Autofac.Pooling.Tests
