+++ test/Autofac.Pooling.Tests/LifetimeScopeTests.cs
@@ -1,6 +1,5 @@
 ﻿using Autofac.Core;
-using Autofac.Pooling.Tests.Types;
-using Autofac.Pooling.Tests.Utils;
+using Autofac.Pooling.Tests.Shared;
 using Xunit;
 
 namespace Autofac.Pooling.Tests
